+++ src/main/java/com/github/games647/securemyaccount/SecureMyAccount.java
@@ -6,14 +6,10 @@ import com.github.games647.securemyaccount.listener.PreventListener;
 import com.github.games647.securemyaccount.listener.SessionListener;
 import com.google.common.collect.Sets;
 
-import java.awt.image.BufferedImage;
 import java.util.Set;
 import java.util.UUID;
 
-import org.bukkit.Bukkit;
 import org.bukkit.entity.Player;
-import org.bukkit.map.MapRenderer;
-import org.bukkit.map.MapView;
 import org.bukkit.plugin.java.JavaPlugin;
 
 public class SecureMyAccount extends JavaPlugin {
@@ -22,10 +18,7 @@ public class SecureMyAccount extends JavaPlugin {
 
     @Override
     public void onEnable() {
-        if (!getDataFolder().exists()) {
-            //create the basic folder
-            getDataFolder().mkdir();
-        }
+        saveDefaultConfig();
 
         //register commands
         getCommand(getName().toLowerCase()).setExecutor(new CreateKeyCommand(this));
@@ -34,16 +27,6 @@ public class SecureMyAccount extends JavaPlugin {
         //register listeners
         getServer().getPluginManager().registerEvents(new PreventListener(this), this);
         getServer().getPluginManager().registerEvents(new SessionListener(this), this);
-    }
-
-    public MapView installRenderer(Player player, BufferedImage image) {
-        MapView mapView = Bukkit.createMap(player.getWorld());
-        for (MapRenderer mapRenderer : mapView.getRenderers()) {
-            mapView.removeRenderer(mapRenderer);
-        }
-
-        mapView.addRenderer(new ImageRenderer(player, image));
-        return mapView;
     }
 
     public void startSession(Player player) {