+++ src/main/java/com/github/games647/securemyaccount/listener/SessionListener.java
@@ -2,13 +2,10 @@ package com.github.games647.securemyaccount.listener;
 
 import com.github.games647.securemyaccount.SecureMyAccount;
 
-import java.util.logging.Level;
-
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
+import org.bukkit.event.player.PlayerJoinEvent;
 import org.bukkit.event.player.PlayerQuitEvent;
-import org.bukkit.event.server.MapInitializeEvent;
-import org.bukkit.map.MapView;
 
 public class SessionListener implements Listener {
 
@@ -19,9 +16,8 @@ public class SessionListener implements Listener {
     }
 
     @EventHandler(ignoreCancelled = true)
-    public void onMap(MapInitializeEvent mapInitializeEvent) {
-        MapView map = mapInitializeEvent.getMap();
-        plugin.getLogger().log(Level.FINE, "Created a map with id={0}", map.getId());
+    public void onPlayerJoin(PlayerJoinEvent joinEvent) {
+
     }
 
     @EventHandler