+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 #Changelog
 
+##0.2
+
+* Added config
+* Added server ip config property
+* Added /register alias for a new account
+* Added /createkey alias for a new account
+
 ##0.1
 
 * First release